+++ src/misc/espresso/irred.c
@@ -304,7 +304,7 @@ start:
 
     if (debug & TAUT) {
         printf("UNATE_REDUCTION: %d unate variables, reduced to %d\n",
-        cdata.vars_unate, CUBELISTSIZE(T));
+        (int)cdata.vars_unate, (int)CUBELISTSIZE(T));
     }
     goto start;
 
@@ -430,7 +430,7 @@ start:
 
     if (debug & TAUT) {
         printf("UNATE_REDUCTION: %d unate variables, reduced to %d\n",
-        cdata.vars_unate, CUBELISTSIZE(T));
+        (int)cdata.vars_unate, (int)CUBELISTSIZE(T));
     }
     goto start;
     }
